use a ad hoc solution to stream binary file. Other disabled

+++ backend/srs/converters/ProtoToStructConverter.hpp
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <cstdint>
 #include <srs/data/SRSDataStructs.hpp>
 #include <srs/data/message.pb.h>
 
@@ -34,7 +35,7 @@ namespace srs::process
 
             auto& header = struct_data.header;
             header.frame_counter = proto_header.frame_counter();
-            header.fec_id = proto_header.fec_id();
+            header.fec_id = static_cast<uint8_t>(proto_header.fec_id());
             header.udp_timestamp = proto_header.udp_timestamp();
             header.overflow = proto_header.overflow();
         }
@@ -42,7 +43,7 @@ namespace srs::process
         static void set_marker_data(const proto::Data& proto, StructData& struct_data)
         {
             const auto& proto_marker_data = proto.marker_data();
-            struct_data.marker_data.reserve(proto_marker_data.size());
+            struct_data.marker_data.reserve(static_cast<uint64_t>(proto_marker_data.size()));
 
             for (const auto& proto_marker : proto_marker_data)
             {
@@ -55,21 +56,21 @@ namespace srs::process
         static void set_hit_data(const proto::Data& proto, StructData& struct_data)
         {
             const auto& proto_hit_data = proto.hit_data();
-            struct_data.hit_data.reserve(proto_hit_data.size());
+            struct_data.hit_data.reserve(static_cast<uint64_t>(proto_hit_data.size()));
 
             for (const auto& proto_hit : proto_hit_data)
             {
                 auto& hit_data = struct_data.hit_data.emplace_back();
 
                 hit_data.is_over_threshold = proto_hit.is_over_threshold();
-                hit_data.channel_num = proto_hit.channel_num();
-                hit_data.tdc = proto_hit.tdc();
-                hit_data.offset = proto_hit.offset();
-                hit_data.vmm_id = proto_hit.vmm_id();
-                hit_data.adc = proto_hit.adc();
-                hit_data.bc_id = proto_hit.bc_id();
+                hit_data.channel_num = static_cast<uint8_t>(proto_hit.channel_num());
+                hit_data.tdc = static_cast<uint8_t>(proto_hit.tdc());
+                hit_data.offset = static_cast<uint8_t>(proto_hit.offset());
+                hit_data.vmm_id = static_cast<uint8_t>(proto_hit.vmm_id());
+                hit_data.adc = static_cast<uint8_t>(proto_hit.adc());
+                hit_data.bc_id = static_cast<uint16_t>(proto_hit.bc_id());
             }
         }
     };
 
-} // namespace srs
+} // namespace srs::process
